# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hi2 Temp Mail API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EmailAddress {
  /** The generated temporary email address */
  email: string;
  /** The username part of the email */
  username: string;
  /** The domain part of the email */
  domain: string;
  /**
     * ISO timestamp when this address expires (if applicable)
     * @nullable
     */
  expiresAt: string | null;
}

export interface EmailMessage {
  /** Unique message identifier */
  id: string;
  /** Sender email address */
  from: string;
  /**
     * Sender display name
     * @nullable
     */
  fromName: string | null;
  /** Email subject line */
  subject: string;
  /** ISO timestamp when message was received */
  receivedAt: string;
  /** Whether the message has been read */
  isRead: boolean;
  /** Short preview of the email body */
  preview: string;
}

export interface Attachment {
  filename: string;
  size: number;
  contentType: string;
}

export interface EmailMessageDetail {
  id: string;
  from: string;
  /** @nullable */
  fromName: string | null;
  subject: string;
  receivedAt: string;
  /**
     * Full HTML body of the email
     * @nullable
     */
  bodyHtml: string | null;
  /**
     * Plain text body of the email
     * @nullable
     */
  bodyText: string | null;
  attachments: Attachment[];
}

export interface MailStats {
  /** Total number of email addresses generated */
  totalGenerated: number;
  /** Number of currently active addresses */
  activeAddresses: number;
  /** Total messages received across all addresses */
  totalMessagesReceived: number;
}

export type GetInboxParams = {
/**
 * The temporary email address to fetch inbox for
 */
email: string;
};

export type FetchMessageParams = {
/**
 * The message ID
 */
messageId: string;
/**
 * The email address the message belongs to
 */
email: string;
};