# lib/api-zod/src/generated/types/attachment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hi2 Temp Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface Attachment {
  filename: string;
  size: number;
  contentType: string;
}