# lib/api-zod/src/generated/types/emailAddress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hi2 Temp Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailAddress {
  /** The generated temporary email address */
  email: string;
  /** The username part of the email */
  username: string;
  /** The domain part of the email */
  domain: string;
  /**
     * ISO timestamp when this address expires (if applicable)
     * @nullable
     */
  expiresAt: string | null;
}