# lib/api-zod/src/generated/types/emailMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hi2 Temp Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailMessage {
  /** Unique message identifier */
  id: string;
  /** Sender email address */
  from: string;
  /**
     * Sender display name
     * @nullable
     */
  fromName: string | null;
  /** Email subject line */
  subject: string;
  /** ISO timestamp when message was received */
  receivedAt: string;
  /** Whether the message has been read */
  isRead: boolean;
  /** Short preview of the email body */
  preview: string;
}