# lib/api-zod/src/generated/types/emailMessageDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hi2 Temp Mail API
 * OpenAPI spec version: 0.1.0
 */
import type { Attachment } from './attachment';

export interface EmailMessageDetail {
  id: string;
  from: string;
  /** @nullable */
  fromName: string | null;
  subject: string;
  receivedAt: string;
  /**
     * Full HTML body of the email
     * @nullable
     */
  bodyHtml: string | null;
  /**
     * Plain text body of the email
     * @nullable
     */
  bodyText: string | null;
  attachments: Attachment[];
}