# lib/api-zod/src/generated/types/mailStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hi2 Temp Mail API
 * OpenAPI spec version: 0.1.0
 */

export interface MailStats {
  /** Total number of email addresses generated */
  totalGenerated: number;
  /** Number of currently active addresses */
  activeAddresses: number;
  /** Total messages received across all addresses */
  totalMessagesReceived: number;
}